move to sepperat clock menu
